generate-dashboard: move matcher to lib file
