+++ .github/actions/generate-dashboard/src/lib.ts
@@ -7,6 +7,7 @@ type ResultLine = {
   outcome: string;
   error: string;
 };
+
 export type ResultFile = ResultLine[];
 
 export const load = (path: string): ResultFile => {
@@ -19,4 +20,20 @@ export const load = (path: string): ResultFile => {
 
 export const save = (path: string, content: string) => {
   fs.writeFileSync(path, content);
+};
+
+type PairOfImplementation = [string, string];
+
+// `     something x something-else    more info ignore`
+const TASK_ID_MATCHER = /\s*(\S+)\s+x\s+(\S+)\s*.*/;
+
+export const fromTaskIdToCoordinate = (
+  taskId: string
+): PairOfImplementation => {
+  const match = taskId.match(TASK_ID_MATCHER);
+  if (!match) {
+    throw new Error(`Task ID ${taskId} does not match the expected format`);
+  }
+
+  return [match[1], match[2]];
 };
